fixed main controller nits
